Fix nginx-proxy custom image to include extra hosts at end
Previously, an include directive for additional hosts used the conf.d/
folder. However, configurations from this folder may be loaded before
the dockergen-generated configuration file and thus definitions made in
that file may not be available to files in the conf.d/ folder.

With this commit, the include directive for the sites-enabled folder is
added to the nginx.tmpl file instead, ensuring that definitions for
additional hosts are loaded only after the dockergen template. This
enables additional hosts to re-use e.g. the vhost log format.